# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel REMS",
        "version": "1.0.1"
    },
    "paths": {
...
        "/api/v1/clients/{id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "clients",
                "description": "clients here",
                "operationId": "clients",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {},
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Delete Client",
                "description": "Delete Client details",
                "operationId": "DeleteClient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Clients"
                ],
                "summary": "Update Client",
                "description": "Update Client details",
                "operationId": "UpdateClient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "phone",
                                    "first_name",
                                    "last_name",
                                    "address"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "4384738483"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "pyaephyo@gmail.com"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "Pyae Phyo"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Khant"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Mandalay"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Agent updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/properties": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get list of properties",
                "operationId": "d00d16f370a4f428d8b8613bc27c7a1a",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Filter properties by city",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "property_type",
                        "in": "query",
                        "description": "Filter properties by property type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Filter properties by minimum price",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Filter properties by maximum price",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "number_of_bedrooms",
                        "in": "query",
                        "description": "Filter properties by number of bedrooms",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "number_of_bathrooms",
                        "in": "query",
                        "description": "Filter properties by number of bathrooms",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort properties by field (price, agent_id)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "price",
                                "agent_id"
                            ]
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Order direction (asc, desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Properties retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Properties retrieved successfully"
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "datas": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Properties"
                ],
                "summary": "properties create",
                "description": "properties create here",
                "operationId": "properties",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "agent_id",
                                    "address",
                                    "city",
                                    "state",
                                    "zip_code",
                                    "property_type",
                                    "price",
                                    "size",
                                    "number_of_bedrooms",
                                    "number_of_bathrooms",
                                    "year_built",
                                    "description",
                                    "status",
                                    "availiablity_type",
                                    "minrental_period",
                                    "approvedby"
                                ],
                                "properties": {
                                    "agent_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Kume"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Kume"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "success"
                                    },
                                    "zip_code": {
                                        "type": "string",
                                        "example": "pp3434"
                                    },
                                    "property_type": {
                                        "type": "string",
                                        "example": "Khant"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": "5000"
                                    },
                                    "size": {
                                        "type": "string",
                                        "example": "2"
                                    },
                                    "number_of_bedrooms": {
                                        "type": "integer",
                                        "example": "5"
                                    },
                                    "number_of_bathrooms": {
                                        "type": "integer",
                                        "example": "5"
                                    },
                                    "year_built": {
                                        "type": "integer",
                                        "example": "5"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "description"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "success"
                                    },
                                    "availiablity_type": {
                                        "type": "string",
                                        "example": "ppk344324"
                                    },
                                    "minrental_period": {
                                        "type": "integer",
                                        "example": "3"
                                    },
                                    "approvedby": {
                                        "type": "string",
                                        "example": "ppk344324"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Agent updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/transaction/{id}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Transactions",
                "description": "Transactions here",
                "operationId": "transaction show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}